third_party/WebKit/LayoutTests/fast/forms/select/option-relayout-modified-text-expected.txt

Modifying text content of an


PASS refSelect.offsetWidth is > testSelect.offsetWidth
PASS testSelect.offsetWidth is refSelect.offsetWidth
PASS successfullyParsed is true

TEST COMPLETE
Pass if the two selects below have the same width.

 
